fixed cow alien, fixed score

+++ Assets/Scripts/UI/Scores.cs
@@ -14,6 +14,6 @@ public class Scores : MonoBehaviour {
 	
 	void Update () {
         highestSpeed = Mathf.Max(highestSpeed, GameObject.FindGameObjectWithTag("Player").GetComponent<MovementManager>().Speed);
-		scoreText.text = "" + highestSpeed;
+		scoreText.text = "" + (int)highestSpeed;
 	}
 }
